implement fuzzy matching for venue finding
This implements a new option, venue_name_field. This is currently
expected to be a node_reference type field (though could be expanded
to be a string in the future). If it's a node_reference, the title of
that node_reference will be used to fuzzy match against the
/3/:url/venues meetup API, which returns a list of the most recently
used venues by the meetup.

As long as your names are basically synchronized, this works great. If
you ever hit a mismatch, create a new venue in meetup named the same
thing as in drupal.